requirements.txt
# KSF V2 Requirements - 简化实用版本
# 专为RTX 5090优化，使用可用的包版本

# 核心深度学习框架
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0

# 数据处理
datasets>=2.10.0
tokenizers>=0.13.0
numpy>=1.21.0
pandas>=1.3.0

# 配置管理
pyyaml>=6.0
omegaconf>=2.3.0

# 训练工具
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0

# 模型保存和加载
safetensors>=0.3.0
huggingface-hub>=0.15.0

# 系统工具
psutil>=5.9.0
rich>=13.0.0

# 可选的高性能库（如果安装失败会跳过）
# flash-attn>=2.0.0
# bitsandbytes>=0.41.0
# deepspeed>=0.9.0 